--- api/bankchurn-api/Procfile
@@ -1,1 +1,0 @@
-web: uvicorn app.main:app --host 0.0.0.0 --port $PORT
