--- api/bankchurn-api/app/schemas/__init__.py
@@ -1,2 +1,0 @@
-from .health import Health
-from .predict import MultipleDataInputs, PredictionResults
