--- api/bankchurn-api/app/schemas/health.py
@@ -1,7 +1,0 @@
-from pydantic import BaseModel
-
-
-class Health(BaseModel):
-    name: str
-    api_version: str
-    model_version: str
