--- api/bankchurn-api/mypy.ini
@@ -1,4 +1,0 @@
-[mypy]
-plugins = pydantic.mypy
-ignore_missing_imports = True
-disallow_untyped_defs = True
