--- api/bankchurn-api/run.sh
@@ -1,1 +1,0 @@
-uvicorn app.main:app --host 0.0.0.0 --port $PORT
