+++ api/flights-delay-api/app/schemas/__init__.py
@@ -1,0 +1,2 @@
+from .health import Health
+from .predict import MultipleDataInputs, PredictionResults
