+++ api/flights-delay-api/app/schemas/health.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel
+
+
+class Health(BaseModel):
+    name: str
+    api_version: str
+    model_version: str
