+++ api/flights-delay-api/mypy.ini
@@ -1,0 +1,4 @@
+[mypy]
+plugins = pydantic.mypy
+ignore_missing_imports = True
+disallow_untyped_defs = True
