+++ api/flights-delay-api/requirements.txt
@@ -1,0 +1,7 @@
+./model-pkg/model_flights_delays-0.0.1-py3-none-any.whl
+uvicorn>=0.20.0,<0.30.0
+fastapi>=0.88.0,<1.0.0
+python-multipart>=0.0.5,<0.1.0
+typing_extensions>=4.2.0,<5.0.0
+loguru>=0.5.3,<1.0.0
+pydantic>=1.10.0,<2.0.0
