+++ api/flights-delay-api/run.sh
@@ -1,0 +1,1 @@
+uvicorn app.main:app --host 0.0.0.0 --port $PORT
