+++ api/flights-delay-api/test_requirements.txt
@@ -1,0 +1,6 @@
+-r requirements.txt
+
+# testing requirements
+pytest>=7.2.0,<8.0.0
+requests>=2.28.0,<2.50.0
+httpx>=0.23.2,<0.50.0
