+++ api/flights-delay-api/typing_requirements.txt
@@ -1,0 +1,6 @@
+# repo maintenance tooling
+black>=22.12.0,<23.0.0
+flake8>=6.0.0,<7.0.0
+mypy>=0.991,<1.0.0
+isort>=5.11.4,<6.0.0
+pydantic>=1.10.4,<1.12.0
